Removing dephy_legacy and adding DephyLegacyActuator class to dephy.
